Improve reconciliation status when commit has pipeline which is not created yet
There is a shortcoming in the GitLab API that sometimes for a few seconds the status and last pipeline
of a commit are None while a pipeline is being created - this is the exact same situation as if
there is no pipeline at all.
Therefore, this change implements some checks to clarify if there is potentially a pipeline,
if yes it waits some timeout for the pipeline to be created. However, because of the dynamic nature of pipelines (rules)
there may not be any pipeline - so the timeout is not perfect ...